requirements.txt
torch==2.3.0
torchvision==0.18.0
numpy==1.26.4
Pillow==10.3.0
tqdm==4.66.4
argparse==1.1